fix(dma-mapping): fix signed size_t error ob build
Replaces "%zd" with "%zu" in a print statement within dma-mapping.c.
The print statment only displays info or debugging so it could be
removed, but it is better to keep it.

`size_t` is almost always unsigned but this statement wanted a signed
`size_t`. No clue why. Turns out %zd is printing as a signed decimal
and %zu is an unsigned decimal, so I just swapped them out in the code.

Source: https://stackoverflow.com/a/2524675/11091039