Another "make even" pull? (#28)
* Remove trace_printk() from kernel source (fix #13)

* Fix broken deploy after name change

* perf(trace_printk): removes traces of trace_printk

* chore(bootimage): disable SELinux for now

* chore(buildscripts): make unpack/repact more explicit [ci skip]

* chore(do-all.sh): Fix "unary operator expected" in BASH

* chore(name scheme): clarify that these are Quanitfy kernels

* fix(SELinux): disable SELinux enforcing in Makefile

* fix(#21): rearrange driver makefile to fix i2c/clk probing (#24)

Fixes #21 by moving `clk` to top of makefile, as suggested by https://stackoverflow.com/questions/30166717/how-to-fix-error-on-clk-getcore-clk-during-probing-driver-i2c-msm-v2

* feat(nfc): add PN544 drivers to defconfig

* fix(dma-mapping): fix signed size_t error ob build

Replaces "%zd" with "%zu" in a print statement within dma-mapping.c.
The print statment only displays info or debugging so it could be
removed, but it is better to keep it.

`size_t` is almost always unsigned but this statement wanted a signed
`size_t`. No clue why. Turns out %zd is printing as a signed decimal
and %zu is an unsigned decimal, so I just swapped them out in the code.

Source: https://stackoverflow.com/a/2524675/11091039

* Revert "fix(dma-mapping): fix signed size_t error ob build"

This reverts commit 3d5284999fd083cb0537214aa7508399ca007041.

* Revert "feat(nfc): add PN544 drivers to defconfig"

This reverts commit 219d3dcbda90309ceedd864dc0e7c9542ea9c9c5.

* feat(PN544 drivers): add PN544 NFC chip drivers

And it works this time...

* fix(unneeded file): remove kernel/out/.config

We don't need this and the file should have been ignored as shown
in the .gitignore.

* fix(defconfig): add missing nfc config entries

* feat(ci): Initial support for ci releases